Refactor. Move memoization functions to a cache.rs file
