+++ src/plots/scatterplot.rs
@@ -19,7 +19,7 @@ use crate::two::shapes::{
     TwoCircle, TwoElement, TwoGroup, TwoLine, TwoPath, TwoRectangle, TwoText,
 };
 
-use crate::render::get_or_init_buffer;
+use crate::cache::get_or_init_buffer;
 
 #[derive(ShaderType, Debug)]
 struct ScatterplotUniforms {
@@ -77,7 +77,8 @@ pub async fn render_scatterplot(
     let labels_bytes: &[u8] = bytemuck::cast_slice(&labels_i32);
 
     // TODO: improve the keys / memoization dependencies to at least include the plot_id and store_name.
-    let x_f32_future = get_or_init_buffer("x_bytes", async || {
+    let x_f32_future_deps = vec!["x_bytes".to_string(), context.params.store_name.to_string(), context.params.plot_id.to_string()];
+    let x_f32_future = get_or_init_buffer(async || {
         let x_array_path = &scatterplot_params.x_key.as_ref();
         let x_array_future = zarrs::array::Array::async_open(store.clone(), x_array_path);
         let x_array = x_array_future.await.unwrap();
@@ -87,9 +88,10 @@ pub async fn render_scatterplot(
         let x_vec = x_result.unwrap();
         let x_f32_inner: Vec<f32> = x_vec.iter().map(|&x| x as f32).collect();
         x_f32_inner
-    });
+    }, &x_f32_future_deps, context.params.cache_enabled);
 
-    let y_f32_future = get_or_init_buffer("y_bytes", async || {
+    let y_f32_future_deps = vec!["y_bytes".to_string(), context.params.store_name.to_string(), context.params.plot_id.to_string()];
+    let y_f32_future = get_or_init_buffer(async || {
         let y_array_path = &scatterplot_params.y_key.as_ref();
         let y_array_future = zarrs::array::Array::async_open(store.clone(), y_array_path);
         let y_array = y_array_future.await.unwrap();
@@ -99,7 +101,7 @@ pub async fn render_scatterplot(
         let y_vec = y_result.unwrap();
         let y_f32_inner: Vec<f32> = y_vec.iter().map(|&y| y as f32).collect();
         y_f32_inner
-    });
+    }, &y_f32_future_deps, context.params.cache_enabled);
 
     // Await in parallel.
     let (x_f32, y_f32) = futures::join!(x_f32_future, y_f32_future);
